requirements.txt
attrs==19.1.0
beautifulsoup4==4.7.1
blis==0.2.4
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
Click==7.0
cymem==2.0.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.2
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-mysql==3.1.0
djangorestframework==3.9.4
dropbox==9.4.0
gTTS==2.0.3
gTTS-token==1.1.3
gunicorn==19.9.0
idna==2.8
jsonschema==3.0.1
mathparse==0.1.2
murmurhash==1.0.2
nltk==3.4.3
numpy==1.16.4
oauthlib==3.0.1
Pillow==6.0.0
Pint==0.9
plac==0.9.6
playsound==1.2.2
preshed==2.0.1
psycopg2==2.8.2
PyJWT==1.7.1
pymongo==3.8.0
pyrsistent==0.15.2
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2019.1
PyYAML==3.13
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.2
spacy==2.1.4
SpeechRecognition==3.8.1
SQLAlchemy==1.2.19
sqlparse==0.3.0
srsly==0.0.7
thinc==7.0.4
tqdm==4.32.2
urllib3==1.25.3
wasabi==0.2.2
whitenoise==4.1.2
